requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.5.2
python-dotenv==0.13.0
python-telegram-bot==12.7
six==1.14.0
toml==0.10.1
tornado==6.0.4
wrapt==1.12.1
